requirements.txt
# Python dependencies for Robot Weed Detector

# Serial communication with Arduino and GPS
pyserial>=3.5

# Optional: Computer vision for weed detection
# Uncomment when implementing camera-based detection
# opencv-python>=4.5.0
# numpy>=1.21.0

# Optional: Machine learning for weed identification
# Uncomment when implementing ML-based classification
# tensorflow>=2.8.0
# or
# torch>=1.10.0
# torchvision>=0.11.0

# Optional: Additional GPS parsing
# pynmea2>=1.18.0
